LMDB: Fix unlock times and add_block functions

+++ src/blockchain_db/blockchain_db.h
@@ -400,7 +400,7 @@ private:
    * @param blk_hash the hash of the block
    */
   virtual void add_block( const block& blk
-                , uint64_t block_weight
+                , size_t block_weight
                 , uint64_t long_term_block_weight
                 , const difficulty_type& cumulative_difficulty
                 , const uint64_t& coins_generated
